+++ lib/src/config/web/js_interop_util.dart
@@ -1,0 +1,1 @@
+export 'js_util_stub.dart' if (dart.library.js_util) 'js_util_web.dart';