+++ lib/src/config/web/js_util_stub.dart
@@ -1,0 +1,8 @@
+dynamic jsify(Object? object) =>
+    throw UnsupportedError('Only supported on web');
+
+Future<T> promiseToFuture<T>(Object promise) =>
+    throw UnsupportedError('Only supported on web');
+
+F allowInterop<F extends Function>(F f) =>
+    throw UnsupportedError('Only supported on web');