+++ lib/src/config/web/js_util_web.dart
@@ -1,0 +1,4 @@
+// ignore_for_file: uri_does_not_exist
+
+// This file is only imported on web platforms where dart:js_util exists
+export 'dart:js_util' show jsify, promiseToFuture, allowInterop;